バックエンドモジュールの構造を文書化し、Rustバックエンドの機能別モジュールを追加。動画ファイルの選択、保存先ディレクトリの選択、ドキュメント保存機能を実装。Gemini APIとの連携を強化し、動画処理機能を追加。新しい型定義を含むtypes.rsを作成し、コードの可読性と保守性を向上させるためにモジュールを分割。
