+++ doc/spec.md
@@ -48,6 +48,63 @@
 - **対応動画形式**:
     - `mp4`, `mpeg`, `mov`, `avi`, `x-flv`, `mpg`, `webm`, `wmv`, `3gpp`, `mkv`
 
+### 3.1. バックエンドモジュール構造
+
+Rustバックエンドは機能別にモジュール化され、保守性と可読性を向上させている。
+
+#### 3.1.1. モジュール一覧
+
+- **`lib.rs`** (391行) - メインライブラリファイル
+  - Tauriコマンドの定義とアプリケーション設定
+  - モジュール間の調整とイベント処理
+  - アプリケーション起動時の初期化処理
+
+- **`types.rs`** (71行) - 型定義モジュール
+  - `VideoFile` - 動画ファイル情報の構造体
+  - `AppSettings` - アプリケーション設定の構造体
+  - `ProgressUpdate` - 進捗更新情報の構造体
+  - `Gemini*` - Gemini API関連の型定義群
+
+- **`file.rs`** (90行) - ファイル操作モジュール
+  - `select_video_files()` - 動画ファイル選択ダイアログ
+  - `select_save_directory()` - 保存先ディレクトリ選択ダイアログ
+  - `save_document_to_file()` - ドキュメントファイル保存処理
+
+- **`gemini.rs`** (509行) - Gemini API連携モジュール
+  - `upload_to_gemini_with_progress()` - 進捗付きファイルアップロード
+  - `generate_with_gemini_with_progress()` - 進捗付きドキュメント生成
+  - `integrate_documents()` - 複数ドキュメントの統合処理
+  - `get_mime_type()` - ファイル形式判定ヘルパー
+
+- **`video.rs`** (213行) - 動画処理モジュール
+  - `split_video_if_needed()` - 長時間動画の自動分割（1時間超の場合）
+  - `get_video_duration()` - ffprobeによる動画時間取得
+  - `check_video_compatibility()` - 動画ファイル互換性チェック
+  - `extract_video_info_with_ffprobe()` - 詳細メタデータ抽出
+  - `get_video_metadata()` - 基本メタデータ取得
+  - `VideoMetadata` - 動画メタデータ構造体
+
+#### 3.1.2. モジュール間の依存関係
+
+```
+lib.rs
+├── types.rs (型定義を参照)
+├── file.rs (ファイル操作を呼び出し)
+├── gemini.rs (API処理を呼び出し)
+└── video.rs (動画処理を呼び出し)
+
+gemini.rs → types.rs (型定義を参照)
+file.rs → types.rs (VideoFile型を参照)
+video.rs → 独立モジュール
+```
+
+#### 3.1.3. 設計原則
+
+- **単一責任の原則**: 各モジュールは特定の機能領域のみを担当
+- **疎結合**: モジュール間の依存を最小限に抑制
+- **型安全性**: 共通の型定義により型安全性を確保
+- **エラーハンドリング**: anyhow::Resultによる統一的なエラー処理
+
 ## 4. 画面構成（案）
 
 - **メイン画面**:
